+++ lib/Ast.ml
@@ -105,6 +105,12 @@ module Ext = struct
   end
 end
 
+module Ext_attrs = struct
+  let has_doc ea =
+    List.exists ~f:Attr.is_doc ea.attrs_before
+    || List.exists ~f:Attr.is_doc ea.attrs_after
+end
+
 module Exp = struct
   let location x = x.pexp_loc
 
@@ -318,27 +324,30 @@ module Structure_item = struct
   let has_doc itm =
     match itm.pstr_desc with
     | Pstr_attribute atr -> Attr.is_doc atr
+    (* one attribute list *)
     | Pstr_eval (_, atrs)
      |Pstr_value {pvbs_bindings= {pvb_attributes= atrs; _} :: _; _}
      |Pstr_primitive {pval_attributes= atrs; _}
      |Pstr_type (_, {ptype_attributes= atrs; _} :: _)
      |Pstr_typext {ptyext_attributes= atrs; _}
      |Pstr_recmodule ({pmb_expr= {pmod_attributes= atrs; _}; _} :: _)
-     |Pstr_modtype {pmtd_attributes= atrs; _}
      |Pstr_open {popen_attributes= atrs; _}
      |Pstr_extension (_, atrs)
      |Pstr_class_type ({pci_attributes= atrs; _} :: _)
      |Pstr_class ({pci_attributes= atrs; _} :: _) ->
         List.exists ~f:Attr.is_doc atrs
-    | Pstr_include
-        {pincl_mod= {pmod_attributes= atrs1; _}; pincl_attributes= atrs2; _}
-     |Pstr_exception
+    | Pstr_exception
         { ptyexn_attributes= atrs1
         ; ptyexn_constructor= {pext_attributes= atrs2; _}
         ; _ }
-     |Pstr_module
-        {pmb_attributes= atrs1; pmb_expr= {pmod_attributes= atrs2; _}; _} ->
+     |Pstr_include
+        {pincl_mod= {pmod_attributes= atrs1; _}; pincl_attributes= atrs2; _}
+      ->
         List.exists ~f:Attr.is_doc atrs1 || List.exists ~f:Attr.is_doc atrs2
+    | Pstr_modtype {pmtd_ext_attrs; _} -> Ext_attrs.has_doc pmtd_ext_attrs
+    | Pstr_module {pmb_ext_attrs; pmb_expr= {pmod_attributes; _}; _} ->
+        Ext_attrs.has_doc pmb_ext_attrs
+        || List.exists ~f:Attr.is_doc pmod_attributes
     | Pstr_value {pvbs_bindings= []; _}
      |Pstr_type (_, [])
      |Pstr_recmodule []
@@ -412,30 +421,33 @@ module Signature_item = struct
   let has_doc itm =
     match itm.psig_desc with
     | Psig_attribute atr -> Attr.is_doc atr
+    (* one attribute list *)
     | Psig_value {pval_attributes= atrs; _}
      |Psig_type (_, {ptype_attributes= atrs; _} :: _)
      |Psig_typesubst ({ptype_attributes= atrs; _} :: _)
      |Psig_typext {ptyext_attributes= atrs; _}
-     |Psig_modtype {pmtd_attributes= atrs; _}
-     |Psig_modtypesubst {pmtd_attributes= atrs; _}
-     |Psig_modsubst {pms_attributes= atrs; _}
      |Psig_open {popen_attributes= atrs; _}
      |Psig_extension (_, atrs)
      |Psig_class_type ({pci_attributes= atrs; _} :: _)
      |Psig_class ({pci_attributes= atrs; _} :: _) ->
         List.exists ~f:Attr.is_doc atrs
-    | Psig_recmodule
-        ( {pmd_type= {pmty_attributes= atrs1; _}; pmd_attributes= atrs2; _}
-        :: _ )
-     |Psig_include
+    (* two attribute list *)
+    | Psig_modtype {pmtd_ext_attrs= ea; _}
+     |Psig_modtypesubst {pmtd_ext_attrs= ea; _}
+     |Psig_modsubst {pms_ext_attrs= ea; _} ->
+        Ext_attrs.has_doc ea
+    | Psig_include
         {pincl_mod= {pmty_attributes= atrs1; _}; pincl_attributes= atrs2; _}
      |Psig_exception
         { ptyexn_attributes= atrs1
         ; ptyexn_constructor= {pext_attributes= atrs2; _}
-        ; _ }
-     |Psig_module
-        {pmd_attributes= atrs1; pmd_type= {pmty_attributes= atrs2; _}; _} ->
+        ; _ } ->
         List.exists ~f:Attr.is_doc atrs1 || List.exists ~f:Attr.is_doc atrs2
+    | Psig_recmodule
+        ({pmd_type= {pmty_attributes= atrs; _}; pmd_ext_attrs= ea; _} :: _)
+     |Psig_module {pmd_ext_attrs= ea; pmd_type= {pmty_attributes= atrs; _}; _}
+      ->
+        Ext_attrs.has_doc ea || (List.exists ~f:Attr.is_doc) atrs
     | Psig_type (_, [])
      |Psig_typesubst []
      |Psig_recmodule []
@@ -510,7 +522,7 @@ module Lb = struct
 end
 
 module Mb = struct
-  let has_doc itm = List.exists ~f:Attr.is_doc itm.pmb_attributes
+  let has_doc itm = Ext_attrs.has_doc itm.pmb_ext_attrs
 
   let is_simple (i, (c : Conf.t)) =
     Poly.(c.fmt_opts.module_item_spacing.v = `Compact)
@@ -524,7 +536,7 @@ module Mb = struct
 end
 
 module Md = struct
-  let has_doc itm = List.exists ~f:Attr.is_doc itm.pmd_attributes
+  let has_doc itm = Ext_attrs.has_doc itm.pmd_ext_attrs
 
   let is_simple (i, (c : Conf.t)) =
     Poly.(c.fmt_opts.module_item_spacing.v = `Compact)
@@ -669,6 +681,8 @@ include T
 
 let is_top = function Top -> true | _ -> false
 
+let attrs_of_ext_attrs ea = ea.attrs_before @ ea.attrs_after
+
 let attributes = function
   | Pld _ -> []
   | Typ x -> x.ptyp_attributes
@@ -677,8 +691,8 @@ let attributes = function
   | Pat x -> x.ppat_attributes
   | Exp x -> x.pexp_attributes
   | Lb x -> x.pvb_attributes
-  | Mb x -> x.pmb_attributes
-  | Md x -> x.pmd_attributes
+  | Mb x -> attrs_of_ext_attrs x.pmb_ext_attrs
+  | Md x -> attrs_of_ext_attrs x.pmd_ext_attrs
   | Cl x -> x.pcl_attributes
   | Mty x -> x.pmty_attributes
   | Mod x -> x.pmod_attributes